Lore established: home page now the scene of our bakery with mrToast to help guide you. Can click on easel to see list and add bread
