+++ breadapp/static/styles/base.css
@@ -1,17 +1,20 @@
-#mainContainer {
-    height: 96%;
-    min-height: 85vh;
-    width: 94vw;
-    padding: 15px;
-    margin-top: 70px;
-    margin-left: auto;
-    margin-right: auto;
-    background-color: rgb(241, 241, 241);
-  }
-
-  .listImage{
-    width: 20px;
-    margin-left: 10px;
-    margin-right: 30px;
-    cursor: pointer;
+body {
+  background-color: black;
 }
+
+#mainContainer {
+    margin-top: 70px;
+    background-color: darkgray;
+}
+.backgroundImage{
+  width: 90vw;
+  height: 60vw;
+  margin-left: auto;
+  margin-right: auto;
+  background-repeat: no-repeat;
+  background-size: cover;
+  position: relative;
+    
+}
+
+
