+++ breadapp/templates/bakery_list.html
@@ -1,0 +1,29 @@
+{% extends 'base.html' %}
+{% load static %}
+{% block content %}
+
+<div id="listBackground" class="backgroundImage">
+    <article class="breadList">
+    <h3>Our Current Selection</h3>
+    {% for bread in bread_inventory %}
+        <p>{{bread.number}}. {{bread.name}} from {{bread.region}}  <a href='{% url "breadapp:bread_details" bread.id %}'><img class="listImage" src="{% static 'images/bread.png' %}"></a></p>
+    {% endfor %}
+    </article>
+    <article class="addBreadForm">
+    <h4>Add Bread to Our Collection!</h4>
+    <form class='addBread' action="{% url 'breadapp:bakery_list' %}" method="post">
+        {% csrf_token %}
+        <fieldset>
+            <label for="breadName">Bread: </label>
+            <input id="breadName" name="breadName" type="text" placeholder="name" required>
+        </fieldset>
+        <fieldset>
+            <label for="breadRegion">from: </label>
+            <input id="breadRegion" name="breadRegion" type="text" placeholder="region" required>
+        </fieldset>
+        <input type="submit" value="Add Bread">
+    </form>
+    </article>  
+</div>
+
+{% endblock %}
