+++ breadapp/views/__init__.py
@@ -1,2 +1,3 @@
 from .bakery_home import bakery_home
+from .bakery_list import bakery_list
 from .bread_details import bread_details
